pyproject.toml
[project]
name = "try-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-aiplatform>=1.70.0",
    "langchain>=1.0.7",
    "langchain-anthropic>=1.1.0",
    "langchain-huggingface>=0.1.2",
    "loadenv>=0.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "sentence-transformers>=3.3.1",
]
